requirements.txt
Pillow>=9.0
numpy>=1.24
imageio>=2.20
pytest>=7.0
psd-tools>=1.9
